converts from split to a regular expression
This script now uses a regular expression to grab the text from zpool
status.

This change was prompted by this script failing to run on one of my
servers, when zpool status seemed to return a different amount of
whitespace before the ata- identifier. Not to mention it is probably the
correct way to do this.